[JENKINS-28492] Catch PingFailureAnalyzer failures
There are times when the PingFailureAnalyzer can fail.  Usually this is just when retrieving available PingFailureAnalyzers (throwing a NPE).  If this happens, the channel will not be properly closed.  This plus certain kinds of network/system failures will cause the channel to not be reset on the server side, and future attempts by the agent to reconnnect will fail.